+++ packages/relayer/.env.example
@@ -19,7 +19,6 @@ RELAYER_EMAIL_ADDR=
 RELAYER_HOSTNAME="example.com"
 WEB_SERVER_ADDRESS="127.0.0.1:4500"
 CIRCUITS_DIR_PATH=  #Path to email-wallet/packages/circuits
-SUBGRAPH_URL=https://api.thegraph.com/subgraphs/name/zkemail/email-wallet-v1-sepolia-2
 INPUT_FILES_DIR_PATH=  #Path to email-wallet/packages/relayer/input_files
 EMAIL_TEMPLATES_PATH=  #Path to email templates, e.g. ./packages/relayer/eml_templates/
